// src/interfaces/achievements.ts
export enum ACHIEVEMENT_EVENT {
  LOGIN = 'LOGIN',
  REGISTER = 'REGISTER',
  PLAY_CHESS_GAME = 'PLAY_CHESS_GAME',
  MOVE_CHESS = 'MOVE_CHESS',
  CHESS_CHECK = 'CHESS_CHECK',
  YOUR_CHESS_CHECK = 'YOUR_CHESS_CHECK',
  WIN_CHESS_GAME = 'WIN_CHESS_GAME',
  LOSE_CHESS_GAME = 'LOSE_CHESS_GAME',
  MOVE_CHECKERS = 'MOVE_CHECKERS',
  CHECKERS_QUEEN = 'CHECKERS_QUEEN',
  PLAY_CHECKERS_GAME = 'PLAY_CHECKERS_GAME',
  WIN_CHECKERS_GAME = 'WIN_CHECKERS_GAME',
  LOSE_CHECKERS_GAME = 'LOSE_CHECKERS_GAME',
}